Move datastore::test_util into its own file.
